Python/backend_flask/requirements.txt
flask
flask-cors
flask-expects-json
python-dotenv